#367 create MassEditTable
